requirements.txt
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
neo4j>=5.15.0
asyncpg>=0.29.0
rich>=13.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0,<5.0
click>=8.1.0
